feat(config): :rocket: add devcontainer settings for workspace
VsCode has a really cool adddition called 'devcontainers' that acts very similar to gitpod, in that it allows you to specify an environment needed for the code in a project to be run. This allows 2 main things I think are cool:||1. When users clone this code locally, vscode can automatically open the code in a docker container with the needed env packages.|2. Github has a gitpod alternative called codespaces that uses this 'devcontainer' concept also.||So adding this feature kills 2 birds with 1 stone in my opinion and keeps this project flexible for users longer term by allowing them to play with this anywhere they want.